move HiddenLogOperatorsTest to hidden dir
