+++ flink/v1.12/flink/src/test/java/com/netease/arctic/flink/read/hidden/kafka/LogKafkaPartitionSplitReaderTest.java
@@ -61,7 +61,7 @@ import static com.netease.arctic.flink.util.kafka.KafkaContainerTest.KAFKA_CONTA
 import static com.netease.arctic.flink.util.kafka.KafkaContainerTest.readRecordsBytes;
 import static com.netease.arctic.flink.write.hidden.BaseLogTest.createLogDataDeserialization;
 import static com.netease.arctic.flink.write.hidden.BaseLogTest.userSchema;
-import static com.netease.arctic.flink.write.hidden.kafka.HiddenLogOperatorsTest.createRowData;
+import static com.netease.arctic.flink.write.hidden.HiddenLogOperatorsTest.createRowData;
 import static org.junit.Assert.assertEquals;
 
 public class LogKafkaPartitionSplitReaderTest {
